Caveat for new HER behavior in CL 4.X
In Cumulus 3.7, in the absence of a BGP PMSI path attribute, ingress replication was assumed by FRR/BGPD and thus BUM forwarding entries were made. But now, in 4.X, unless there is an explicit PMSI path attribute that asserts ingress replication in the type-3 advertisement, flooding is disabled. 

This caveat entry calls attention to this change in behavior and is hopefully easily searchable for folks coming here confused about new BUM replication behavior.